requirements.txt
Flask
gunicorn
#apt-get install libyaml-dev
PyYaml
pymongo
